开火延迟改为int

+++ src/main/java/com/atsuishio/superbwarfare/item/gun/data/GunData.java
@@ -277,7 +277,7 @@ public class GunData {
         return defaultGunData().burstAmount;
     }
 
-    public double shootDelay() {
+    public int shootDelay() {
         return defaultGunData().shootDelay;
     }
